+++ ArkMonitor/controllers/Misc/Section Headers/ForgedBlockSectionHeader.swift
@@ -1,5 +1,5 @@
 //
-//  ForgedBlockTableViewCell.swift
+//  ForgedBlockSectionHeader.swift
 //  ArkMonitor
 //
 //  Created by Andrew on 2017-09-12.
@@ -8,24 +8,18 @@
 
 import UIKit
 
-class ForgedBlockTableViewCell: UITableViewCell {
+class ForgedBlockSectionHeader: UIView {
     
-    var heightLabel  : UILabel!
-    var timeLabel    : UILabel!
-    var feeLabel     : UILabel!
-    var rewardLabel  : UILabel!
-    var seperator    : UIView!
-
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
+    override init(frame: CGRect) {
+        super.init(frame: frame)
         
-        backgroundColor = ArkPalette.secondaryBackgroundColor
-        selectionStyle  = .none
+        backgroundColor = ArkPalette.backgroundColor
         
-        heightLabel = UILabel()
+        let heightLabel = UILabel()
         heightLabel.textColor = ArkPalette.textColor
+        heightLabel.text = "Height"
         heightLabel.textAlignment = .center
-        heightLabel.font = UIFont.systemFont(ofSize: 14.0, weight: .light)
+        heightLabel.font = UIFont.systemFont(ofSize: 15.0, weight: .semibold)
         addSubview(heightLabel)
         
         heightLabel.snp.makeConstraints { (make) in
@@ -33,10 +27,11 @@ class ForgedBlockTableViewCell: UITableViewCell {
             make.width.equalToSuperview().multipliedBy(0.3)
         }
         
-        timeLabel = UILabel()
+        let timeLabel = UILabel()
         timeLabel.textColor = ArkPalette.textColor
+        timeLabel.text = "Time"
         timeLabel.textAlignment = .center
-        timeLabel.font = UIFont.systemFont(ofSize: 14.0, weight: .light)
+        timeLabel.font = UIFont.systemFont(ofSize: 15.0, weight: .semibold)
         addSubview(timeLabel)
         
         timeLabel.snp.makeConstraints { (make) in
@@ -45,10 +40,11 @@ class ForgedBlockTableViewCell: UITableViewCell {
             make.width.equalToSuperview().multipliedBy(0.3)
         }
         
-        feeLabel = UILabel()
-        feeLabel.textColor = ArkPalette.highlightedTextColor
+        let feeLabel = UILabel()
+        feeLabel.textColor = ArkPalette.textColor
+        feeLabel.text = "Fee"
         feeLabel.textAlignment = .center
-        feeLabel.font = UIFont.systemFont(ofSize: 14.0, weight: .light)
+        feeLabel.font = UIFont.systemFont(ofSize: 15.0, weight: .semibold)
         addSubview(feeLabel)
         
         feeLabel.snp.makeConstraints { (make) in
@@ -57,10 +53,11 @@ class ForgedBlockTableViewCell: UITableViewCell {
             make.width.equalToSuperview().multipliedBy(0.2)
         }
         
-        rewardLabel = UILabel()
-        rewardLabel.textColor = ArkPalette.highlightedTextColor
+        let rewardLabel = UILabel()
+        rewardLabel.textColor = ArkPalette.textColor
+        rewardLabel.text = "Reward"
         rewardLabel.textAlignment = .center
-        rewardLabel.font = UIFont.systemFont(ofSize: 14.0, weight: .light)
+        rewardLabel.font = UIFont.systemFont(ofSize: 15.0, weight: .semibold)
         addSubview(rewardLabel)
         
         rewardLabel.snp.makeConstraints { (make) in
@@ -69,30 +66,24 @@ class ForgedBlockTableViewCell: UITableViewCell {
             make.width.equalToSuperview().multipliedBy(0.2)
         }
         
-        seperator = UIView()
+        let seperator = UIView()
         seperator.backgroundColor = ArkPalette.tertiaryBackgroundColor
         addSubview(seperator)
         seperator.snp.makeConstraints { (make) in
             make.left.right.bottom.equalToSuperview()
             make.height.equalTo(0.5)
         }
+        
+        let seperator2 = UIView()
+        seperator2.backgroundColor = ArkPalette.tertiaryBackgroundColor
+        addSubview(seperator2)
+        seperator2.snp.makeConstraints { (make) in
+            make.left.right.top.equalToSuperview()
+            make.height.equalTo(0.5)
+        }
     }
     
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-    
-    public func update(_ block: Block) {
-        heightLabel.text = String(block.height)
-        timeLabel.text   = Utils.getTimeAgo(timestamp: Double(block.timestamp))
-        feeLabel.text    = String(Utils.convertToArkBase(value: Int64(block.totalFee)))
-        rewardLabel.text = String(Utils.convertToArkBase(value: Int64(block.reward)))
-        
-        backgroundColor           = ArkPalette.secondaryBackgroundColor
-        heightLabel.textColor     = ArkPalette.textColor
-        timeLabel.textColor       = ArkPalette.textColor
-        feeLabel.textColor        = ArkPalette.highlightedTextColor
-        rewardLabel.textColor     = ArkPalette.highlightedTextColor
-        seperator.backgroundColor = ArkPalette.tertiaryBackgroundColor
     }
 }
